pull down of QueryHandler into NetworkManager
